backend/pyproject.toml
[project]
name = "clevercolab-classifier"
version = "0.1.0"
description = "Logistics document classifier for Chilean customs agencies"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "anthropic>=0.52.0",
    "openai>=1.82.0",
    "google-genai>=1.14.0",
    "httpx>=0.28.0",
    "pymupdf>=1.25.0",
    "mistralai>=1.5.0",
    "boto3>=1.36.0",
    "mangum>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "moto[dynamodb,s3,sqs]>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
